requirements.txt
streamlit==1.20.0
whisper==1.0.0
pyannote.audio==2.1.1
torch==2.0.0
transformers==4.27.0
